requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
pydantic-settings
gspread
google-auth